try fixing tarot variant, try adding belote variant

+++ desktop/applis/apps/core/games/tarot/src/main/java/cards/tarot/RulesTarot.java
@@ -38,7 +38,7 @@ public final class RulesTarot {
         }
     }
 
-    public RulesTarot(byte _nbPlayers) {
+    public RulesTarot(int _nbPlayers) {
         initAllowBids();
         boolean found_ = false;
         for (DealingTarot r: DealingTarot.getRepartitionsValides()) {
